refactor: 삭제 날짜 추가 in 로그
