+++ src/main/java/com/c4cometrue/mystorage/filedeletionlog/FileDeletionLogService.java
@@ -1,4 +1,4 @@
-package com.c4cometrue.mystorage.fileDeletionLog;
+package com.c4cometrue.mystorage.filedeletionlog;
 
 import com.c4cometrue.mystorage.file.FileMetadata;
 import lombok.RequiredArgsConstructor;
@@ -12,13 +12,7 @@ public class FileDeletionLogService {
     private final FileDeletionLogRepository fileDeletionLogRepository;
 
     public void saveFileDeleteLog(List<FileMetadata> files) {
-        List<FileDeletionLog> logs = files.stream()
-                .map(file -> FileDeletionLog.builder()
-                        .originalFileName(file.getOriginalFileName())
-                        .filePath(file.getFilePath())
-                        .deleterId(file.getUploaderId())
-                        .build())
-                .toList();
+        List<FileDeletionLog> logs = files.stream().map(file -> FileDeletionLog.builder().originalFileName(file.getOriginalFileName()).filePath(file.getFilePath()).deleterId(file.getUploaderId()).build()).toList();
 
         fileDeletionLogRepository.saveAll(logs);
     }
