+++ src/test/java/com/c4cometrue/mystorage/fileDeleteLog/FileDeleteLogTest.java
@@ -1,6 +1,6 @@
 package com.c4cometrue.mystorage.fileDeleteLog;
 
-import com.c4cometrue.mystorage.fileDeletionLog.FileDeletionLog;
+import com.c4cometrue.mystorage.filedeletionlog.FileDeletionLog;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -19,7 +19,10 @@ class FileDeleteLogTest {
                 .deleterId(FILE_METADATA.getUploaderId())
                 .build();
 
+        log.prePersist();
+
         Assertions.assertNotNull(log);
+        Assertions.assertNotNull(log.getDeleteAt());
         Assertions.assertEquals(ORIGINAL_FILE_NAME, log.getOriginalFileName());
         Assertions.assertEquals(USER_PATH, log.getFilePath());
         Assertions.assertEquals(USER_ID, log.getDeleterId());
