+++ src/test/java/com/c4cometrue/mystorage/storage/StorageFacadeServiceTest.java
@@ -3,7 +3,7 @@ package com.c4cometrue.mystorage.storage;
 import com.c4cometrue.mystorage.file.FileService;
 import com.c4cometrue.mystorage.file.dto.CursorFileResponse;
 import com.c4cometrue.mystorage.file.dto.FileContent;
-import com.c4cometrue.mystorage.fileDeletionLog.FileDeletionLogService;
+import com.c4cometrue.mystorage.filedeletionlog.FileDeletionLogService;
 import com.c4cometrue.mystorage.folder.FolderService;
 import com.c4cometrue.mystorage.folder.dto.CursorFolderResponse;
 import com.c4cometrue.mystorage.folder.dto.FolderContent;
